MER: Don't attach a NULL buffer when receiving a onscreen_visibility(Hidden) event
We're kinda cheating here, and not doing what the protocol says so I'm not sure
this would make sense upstream. However we really don't want to attach a NULL
buffer here, because we don't wait for surafces to draw a new buffer before
showing them, and we don't want random garbage either.