Prevent QWaylandQuickSurface from holding onto multiple buffers indefinitely.
With a queue of just two buffers BufferAttacher can end up holding
references to both after the renderer is stopped starving the client
application and causing it to block in the glSwapBuffers until the
renderer restarts or the surface is destroyed.  Release the current
buffer to the client when the renderer is stopped so it can continue.

Change-Id: Ica0e13ef78f7e6058e273c26b517a88d07f958c7