[qtwayland] Fix copying to the clipboard failing.  Fixes JB#32781
A ulong timestamp with millisecond resolution overflows after about
49 hours meaning a direct comparison between timestamps is only
valid until that wrap point is reached.

Given the timestamps originate in the compositor when the data source
object is created it is difficult to see what information they provide
which isn't implicit in the order of execution of calls to the function
the check is in, making removing the check the safest course of action.

Change-Id: I1b1d4458b913d54028e5e58d7eb3f217cc5b8186