[wayland] Fix a crash when an application exits. JB#44844
Two things try to delete QWaylandSurfaceInterface instances, the
QWaylandSurface they were created for and the wayland resource they
represent. In my case the QWaylandSurface is doing so first and there
was a subsequent crash due to an access after deletion by the code
generated resource destroy callback which can be avoided by destroying
the resource from within the object destructor and shortcutting the
consquent destruction. The reverse order will still be a problem.